requirements.txt
Flask==2.0.1
crccheck>=0.6
pyserial
pythoncrc
git+https://github.com/LiamBindle/PyVESC
Flask-SocketIO
eventlet
